refactor: http.model 패키지 변경
- common, request, response
