+++ src/main/java/http/model/request/HttpMethod.java
@@ -1,4 +1,4 @@
-package http.model;
+package http.model.request;
 
 import http.exceptions.IllegalHttpRequestException;
