+++ src/main/java/http/model/request/HttpParameters.java
@@ -1,4 +1,4 @@
-package http.model;
+package http.model.request;
 
 import java.util.Collections;
 import java.util.HashMap;
