+++ src/main/java/http/supoort/converter/response/ResponseMessageConverter.java
@@ -1,7 +1,6 @@
-package http.supoort.converter;
+package http.supoort.converter.response;
 
 import http.exceptions.FailToConvertMessageException;
-import http.model.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
