+++ src/main/java/http/view/TempleteEngineResolver.java
@@ -1,7 +1,5 @@
 package http.view;
 
-import http.model.HttpResponse;
-
 public class TempleteEngineResolver implements ViewResolver {
     @Override
     public HttpResponse resolve(ModelAndView modelAndView) {
