+++ src/main/java/http/view/ViewResolver.java
@@ -1,7 +1,5 @@
 package http.view;
 
-import http.model.HttpResponse;
-
 public interface ViewResolver {
     HttpResponse resolve(ModelAndView modelAndView);
 }
