+++ src/test/java/http/model/ContentTypeTest.java
@@ -1,6 +1,7 @@
 package http.model;
 
 import http.exceptions.IllegalHttpRequestException;
+import http.model.response.ContentType;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
