+++ src/test/java/http/model/HttpMethodTest.java
@@ -1,6 +1,7 @@
 package http.model;
 
 import http.exceptions.IllegalHttpRequestException;
+import http.model.request.HttpMethod;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
