+++ src/test/java/http/model/HttpParametersTest.java
@@ -1,5 +1,6 @@
 package http.model;
 
+import http.model.request.HttpParameters;
 import org.junit.jupiter.api.Test;
 
 import java.util.Map;
